requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.18
pydantic==2.10.4
pillow==11.0.0
numpy==2.2.1
fast-alpr[onnx]==0.2.3

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
